// src/app/models/user/UserModal.ts
export class UserModal {
  result?: string;
  userId?: string;
  userName?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  dob?: string;
  address?: string;
  address2?: string;
  city?: string;
  pin?: string;
  state?: string;
  gender?: string;
  email?: string;
  mobile?: string;
  kycVerified?: string;
  registrationDate?: string;
  isFraud?: string;
  fraudReason?: string;
  markedFraudDate?: string;
  isLocked?:string;
  code?: string;
  deposit?: string;
  withdrawable?: string;
  pokerDeposit?: string;
  pokerWithdrawal?: string;
  promoBonus?: string;
  lockedBonus?: string;
  totalCashGamesPlayed?: string;
  totalFreeGamesPlayed?: string;
  totalRake?: string;
  totalDepositAmount?: string;
  totalWithdrawalAmount?: string;
  markedNonFraudDate: string;
  emailVerified?: string;
  mobileVerified?: string;
}