// src/app/models/reports/callBreak/callBreak.ts
export class callBreak {
    gameId: string;
    tableId: string;
    tableName: string;
    entryFee: string;
    rake: string;
    winners: [];
    username: string;
    amount: string;
    status: string;
    startTime: string;
    endTime: string;
    noOfRounds: string;
    initalCardsDetails: [];
    cards: string;
    realPlayer: string
}