// src/app/models/reports/callBreak/roundDetails.ts
export class roundDetails {
    gameId: string;
    tableId: string;
    tableName: string;
    entryFee: string;
    rake: string;
    winners: [];
    username: string;
    amount: string;
    status: string;
    startTime: string;
    endTime: string;
    noOfRounds: string;
    initalCardsDetails: [];
    playedCards: []
    cards: string;
    realPlayer: string
}